passing menu to a file with its name
